pyproject.toml
[tool.poetry]
name = "echo"
version = "0.1.0"
description = "A distributed converstional assistant system for the Monash Smart Manufacturing lab."
authors = ["Tyson Truong <tystruong@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
jupyter = "^1.1.1"
openai = "^1.45.0"
paho-mqtt = "^2.1.0"
prompt-toolkit = "^3.0.47"
pymilvus = {extras = ["model"], version = "^2.4.6"}
ipykernel = "^6.29.5"
pymupdf = "^1.24.10"
langchain = "^0.3.0"
langchain-community = "^0.3.0"
langchain-experimental = "^0.3.0"
langchain-openai = "^0.2.0"
pypdf = "^4.3.1"
google-generativeai = "^0.8.1"
tqdm = "^4.66.5"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
